some shit

+++ SalesServices/SalesServices/Views/AdminPage.xaml.cs
@@ -33,17 +33,24 @@ namespace SalesServices.Views
 
         private void AddProductButton_Click(object sender, RoutedEventArgs e)
         {
-            _viewModel.OpenManagerWindow(new Product());
+            _viewModel.OpenManagerWindow(new Product() { DateOfAdd=DateTime.Now, Discount=1});
+            _viewModel.UpdateProductsList();
         }
 
         private void EditProductButton_Click(object sender, RoutedEventArgs e)
         {
             _viewModel.OpenManagerWindow(_viewModel.SelectedProduct);
+            _viewModel.UpdateProductsList();
         }
 
         private void DeleteProductButton_Click(object sender, RoutedEventArgs e)
         {
-            _viewModel.ProductService.Delete(_viewModel.SelectedProduct);
+            var result = MessageBox.Show($"Вы уверены что хотите удалить {_viewModel.SelectedProduct.Title}?","Внимание!",MessageBoxButton.YesNo,MessageBoxImage.Question);
+            if(result == MessageBoxResult.Yes)
+            {
+                _viewModel.ProductService.Delete(_viewModel.SelectedProduct);
+                _viewModel.UpdateProductsList();
+            }
         }
 
         private void CleanProductButton_Click(object sender, RoutedEventArgs e)
@@ -53,17 +60,24 @@ namespace SalesServices.Views
 
         private void AddServiceButton_Click(object sender, RoutedEventArgs e)
         {
-            _viewModel.OpenManagerWindow(new Service());
+            _viewModel.OpenManagerWindow(new Service() { DateOfAdd = DateTime.Now, CostPerHour=0 });
+            _viewModel.UpdateServicesList();
         }
 
         private void EditServiceButton_Click(object sender, RoutedEventArgs e)
         {
             _viewModel.OpenManagerWindow(_viewModel.SelectedService);
+            _viewModel.UpdateServicesList();
         }
 
         private void DeleteServiceButton_Click(object sender, RoutedEventArgs e)
         {
-            _viewModel.ServiceService.Delete(_viewModel.SelectedService);
+            var result = MessageBox.Show($"Вы уверены что хотите удалить {_viewModel.SelectedService.Title}?", "Внимание!", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (result == MessageBoxResult.Yes)
+            {
+                _viewModel.ServiceService.Delete(_viewModel.SelectedService);
+                _viewModel.UpdateServicesList();
+            }
         }
 
         private void CleanServiceButton_Click(object sender, RoutedEventArgs e)
@@ -74,16 +88,23 @@ namespace SalesServices.Views
         private void AddUserButton_Click(object sender, RoutedEventArgs e)
         {
             _viewModel.OpenManagerWindow(new User());
+            _viewModel.UpdateUsersList();
         }
 
         private void EditUserButton_Click(object sender, RoutedEventArgs e)
         {
             _viewModel.OpenManagerWindow(_viewModel.SelectedUser);
+            _viewModel.UpdateUsersList();
         }
 
         private void DeleteUserButton_Click(object sender, RoutedEventArgs e)
         {
-            _viewModel.UserService.Delete(_viewModel.SelectedUser, _viewModel.SelectedUser.UserProfile);
+            var result = MessageBox.Show($"Вы уверены что хотите удалить {_viewModel.SelectedUser.UserProfile.FullName}?", "Внимание!", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (result == MessageBoxResult.Yes)
+            {
+                _viewModel.UserService.Delete(_viewModel.SelectedUser, _viewModel.SelectedUser.UserProfile);
+                _viewModel.UpdateUsersList();
+            }
         }
 
         private void CleanUserButton_Click(object sender, RoutedEventArgs e)
@@ -93,17 +114,24 @@ namespace SalesServices.Views
 
         private void AddUserProductButton_Click(object sender, RoutedEventArgs e)
         {
-            _viewModel.OpenManagerWindow(new UserProduct());
+            _viewModel.OpenManagerWindow(new UserProduct() { DateOfOrder=DateTime.Now });
+            _viewModel.UpdateUserProductsList();
         }
 
         private void EditUserProductButton_Click(object sender, RoutedEventArgs e)
         {
             _viewModel.OpenManagerWindow(_viewModel.SelectedUserProduct);
+            _viewModel.UpdateUserProductsList();
         }
 
         private void DeleteUserProductButton_Click(object sender, RoutedEventArgs e)
         {
-            _viewModel.UserProductsService.Delete(_viewModel.SelectedUserProduct);
+            var result = MessageBox.Show($"Вы уверены что хотите удалить заказ продукта {_viewModel.SelectedUserProduct.Product.Title} пользователя {_viewModel.SelectedUserProduct.User.UserProfile.FullName}?", "Внимание!", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (result == MessageBoxResult.Yes)
+            {
+                _viewModel.UserProductsService.Delete(_viewModel.SelectedUserProduct);
+                _viewModel.UpdateUserProductsList();
+            }
         }
 
         private void CleanUserProductButton_Click(object sender, RoutedEventArgs e)
@@ -113,17 +141,24 @@ namespace SalesServices.Views
 
         private void AddUserServiceButton_Click(object sender, RoutedEventArgs e)
         {
-            _viewModel.OpenManagerWindow(new UserSvc());
+            _viewModel.OpenManagerWindow(new UserSvc() { DateOfOrder = DateTime.Now });
+            _viewModel.UpdateUserServicesList();
         }
 
         private void EditUserServiceButton_Click(object sender, RoutedEventArgs e)
         {
             _viewModel.OpenManagerWindow(_viewModel.SelectedUserService);
+            _viewModel.UpdateUserServicesList();
         }
 
         private void DeleteUserServiceButton_Click(object sender, RoutedEventArgs e)
         {
-            _viewModel.UserServicesService.Delete(_viewModel.SelectedUserService);
+            var result = MessageBox.Show($"Вы уверены что хотите удалить заказ услуги {_viewModel.SelectedUserProduct.Product.Title} пользователя {_viewModel.SelectedUserProduct.User.UserProfile.FullName}?", "Внимание!", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (result == MessageBoxResult.Yes)
+            {
+                _viewModel.UserServicesService.Delete(_viewModel.SelectedUserService);
+                _viewModel.UpdateUserServicesList();
+            }
         }
 
         private void CleanUserServiceButton_Click(object sender, RoutedEventArgs e)
